% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDE-LAPTOP-HQC740DP.R, R/PDE.R
\name{.PDE_extr_data_from_pdf}
\alias{.PDE_extr_data_from_pdf}
\title{Extracting data from a PDF (Protable Document Format) file}
\usage{
.PDE_extr_data_from_pdf(
  pdf,
  whattoextr,
  out = ".",
  filter.words = "",
  regex.fw = TRUE,
  ignore.case.fw = FALSE,
  filter.word.times = 20,
  table.heading.words = "",
  ignore.case.th = FALSE,
  search.words,
  search.word.categories = NULL,
  regex.sw = TRUE,
  ignore.case.sw = FALSE,
  eval.abbrevs = TRUE,
  out.table.format = ".csv (WINDOWS-1252)",
  dev_x = 20,
  dev_y = 9999,
  context = 0,
  write.table.locations = FALSE,
  exp.nondetc.tabs = TRUE,
  write.tab.doc.file = TRUE,
  write.txt.doc.file = TRUE,
  delete = TRUE,
  cpy_mv = "nocpymv",
  verbose = TRUE
)

.PDE_extr_data_from_pdf(
  pdf,
  whattoextr,
  out = ".",
  filter.words = "",
  regex.fw = TRUE,
  ignore.case.fw = FALSE,
  filter.word.times = 20,
  table.heading.words = "",
  ignore.case.th = FALSE,
  search.words,
  search.word.categories = NULL,
  regex.sw = TRUE,
  ignore.case.sw = FALSE,
  eval.abbrevs = TRUE,
  out.table.format = ".csv (WINDOWS-1252)",
  dev_x = 20,
  dev_y = 9999,
  context = 0,
  write.table.locations = FALSE,
  exp.nondetc.tabs = TRUE,
  write.tab.doc.file = TRUE,
  write.txt.doc.file = TRUE,
  delete = TRUE,
  cpy_mv = "nocpymv",
  verbose = TRUE
)
}
\arguments{
\item{pdf}{String. Path to the PDF file to be analyzed.}

\item{whattoextr}{String. Either \emph{txt}, \emph{tab}, or \emph{tabandtxt}
for PDFS2TXT (extract sentences from a PDF file) or PDFS2TABLE (table of a PDF
file to a Microsoft Excel file) extraction. \emph{tab} allows the extraction
of tables with and without search words while \emph{txt} and \emph{tabandtxt}
require search words.}

\item{out}{String. Directory chosen to save analysis results in. Default:
\code{"."}.}

\item{filter.words}{List of strings. The list of filter words. If not
\code{NA} or \code{""} a hit will be counted every time a word from the list
is detected in the article.
 Default: \code{""}.}

\item{regex.fw}{Logical. If TRUE filter words will follow the regex rules
(see \url{https://github.com/erikstricker/PDE/blob/master/inst/examples/cheetsheets/regex.pdf}).
 Default = \code{TRUE}.}

\item{ignore.case.fw}{Logical. Are the filter words case-sensitive (does
capitalization matter)? Default: \code{FALSE}.}

\item{filter.word.times}{Numeric. The minimum number of hits described for
\code{filter.words} for a paper to be further analyzed. Default: \code{20}.}

\item{table.heading.words}{List of strings. Different than standard (TABLE,
TAB or table plus number) headings to be detected. Regex rules apply (see
also
\url{https://github.com/erikstricker/PDE/blob/master/inst/examples/cheetsheets/regex.pdf}).
 Default = \code{""}.}

\item{ignore.case.th}{Logical. Are the additional table headings (see
\code{table.heading.words}) case-sensitive (does capitalization matter)?
Default = \code{FALSE}.}

\item{search.words}{List of strings. List of search words. To extract all
tables from the PDF file leave \code{search.words = ""}.}

\item{regex.sw}{Logical. If TRUE search words will follow the regex rules
(see \url{https://github.com/erikstricker/PDE/blob/master/inst/examples/cheetsheets/regex.pdf}).
 Default = \code{TRUE}.}

\item{ignore.case.sw}{Logical. Are the search words case-sensitive (does
capitalization matter)? Default: \code{FALSE}.}

\item{eval.abbrevs}{Logical. Should abbreviations for the search words be
automatically detected and then replaced with the search word + "$*"?
Default: \code{TRUE}.}

\item{out.table.format}{String. Output file format. Either comma separated
file \code{.csv} or tab separated file \code{.tsv}. The encoding indicated
in parantheses should be selected according to the operational system 
exported tables are opened in, i.e., Windows: \code{"(WINDOWS-1252)"}; Mac: 
\code{(macintosh)}; Linux: \code{(UTF-8)}. Default: \code{".csv"} and 
encoding depending on the operational system.}

\item{dev_x}{Numeric. For a table the size of indention which would be
considered the same column. Default: \code{20}.}

\item{dev_y}{Numeric. For a table the vertical distance which would be
considered the same row. Can be either a number or set to dynamic detection 
[9999], in which case the font size is used to detect which words are in the 
same row. 
Default: \code{9999}.}

\item{context}{Numeric. Number of sentences extracted before and after the
sentence with the detected search word. If \code{0} only the sentence with
the search word is extracted. Default: \code{0}.}

\item{write.table.locations}{Logical. If \code{TRUE}, a separate file with the
headings of all tables, their relative location in the generated html and
txt files, as well as information if search words were found will be
generated. Default: \code{FALSE}.}

\item{exp.nondetc.tabs}{Logical. If \code{TRUE}, if a table was detected in a
PDF file but is an image or cannot be read, the page with the table with be
exported as a png. Default: \code{TRUE}.}

\item{write.tab.doc.file}{Logical. If \code{TRUE}, if search words are used
for table detection and no search words were found in the tables of a PDF 
file, a \strong{no.table.w.search.words}. Default: \code{TRUE}.}

\item{write.txt.doc.file}{Logical. If \code{TRUE}, if no search words were
found in the sentences of a PDF file, a file will be created with the PDF
filename followed by \strong{no.txt.w.search.words}. If the PDF file is
empty, a file will be created with the PDF filename followed by
\strong{no.content.detected}. If the filter word threshold is not met, 
a file will be created with the PDF filename followed by 
\strong{no.txt.w.filter.words}. Default: \code{TRUE}.}

\item{delete}{Logical. If \code{TRUE}, the intermediate \strong{txt},
\strong{keeplayouttxt} and \strong{html} copies of the PDF file will be 
deleted. Default: \code{TRUE}.}

\item{cpy_mv}{String. Either "nocpymv", "cpy", or "mv". If filter words are used in the
analyses, the processed PDF files will either be copied ("cpy") or moved ("mv") into the
/pdf/ subfolder of the output folder. Default: \code{"nocpymv"}.}

\item{verbose}{Logical. Indicates whether messages will be printed in the 
console. Default: \code{TRUE}.}
}
\value{
If tables were extracted from the PDF file the function returns a list of
 following tables/items: 1) \strong{htmltablelines}, 2)
 \strong{txttablelines}, 3) \strong{keeplayouttxttablelines}, 4) \strong{id},
 5) \strong{out_msg}.
 The \strong{tablelines} are tables that provide the heading and position of
 the detected tables. The \strong{id} provide the name of the PDF file. The
 \strong{out_msg} includes all messages printed to the console or the suppressed
 messages if \code{verbose=FALSE}.

If tables were extracted from the PDF file the function returns a list of
 following tables/items: 1) \strong{htmltablelines}, 2)
 \strong{txttablelines}, 3) \strong{keeplayouttxttablelines}, 4) \strong{id},
 5) \strong{out_msg}.
 The \strong{tablelines} are tables that provide the heading and position of
 the detected tables. The \strong{id} provide the name of the PDF file. The
 \strong{out_msg} includes all messages printed to the console or the suppressed
 messages if \code{verbose=FALSE}.
}
\description{
\code{PDE_extr_data_from_pdf} extracts sentences or tables from a single PDF
file and writes output in the corresponding folder.

\code{PDE_extr_data_from_pdf} extracts sentences or tables from a single PDF
file and writes output in the corresponding folder.
}
\examples{

## Running a simple analysis with filter and search words to extract sentences and tables
if(PDE_check_Xpdf_install() == TRUE){
 outputtables <- .PDE_extr_data_from_pdf(pdf = "/examples/Methotrexate/29973177_!.pdf",
 whattoextr = "tabandtxt",
 out = paste0(system.file(package = "PDE"),"/examples/MTX_output+-0_test/"),
 filter.words = strsplit("cohort;case-control;group;study population;study participants", ";")[[1]],
 ignore.case.fw = TRUE,
 regex.fw = FALSE,
 search.words = strsplit("(M|m)ethotrexate;(T|t)rexal;(R|r)heumatrex;(O|o)trexup", ";")[[1]],
 ignore.case.sw = FALSE,
 regex.sw = TRUE)
}

## Running an advanced analysis with filter and search words to
## extract sentences and tables and obtain documentation files
if(PDE_check_Xpdf_install() == TRUE){
 outputtables <- .PDE_extr_data_from_pdf(pdf = paste0(system.file(package = "PDE"),
                       "/examples/Methotrexate/29973177_!.pdf"),
 whattoextr = "tabandtxt",
 out = paste0(system.file(package = "PDE"),"/examples/MTX_output+-1_test/"),
 context = 1,
 dev_x = 20,
 dev_y = 9999,
 filter.words = strsplit("cohort;case-control;group;study population;study participants", ";")[[1]],
 ignore.case.fw = TRUE,
 regex.fw = FALSE,
 filter.word.times = 20,
 table.heading.words = "",
 ignore.case.th = FALSE,
 search.words = strsplit("(M|m)ethotrexate;(T|t)rexal;(R|r)heumatrex;(O|o)trexup", ";")[[1]],
 ignore.case.sw = FALSE,
 regex.sw = TRUE,
 eval.abbrevs = TRUE,
 out.table.format = ".csv (WINDOWS-1252)",
 write.table.locations = TRUE,
 write.tab.doc.file = TRUE,
 write.txt.doc.file = TRUE,
 exp.nondetc.tabs = TRUE,
 cpy_mv = "nocpymv",
 delete = TRUE)
}


## Running a simple analysis with filter and search words to extract sentences and tables
if(PDE_check_Xpdf_install() == TRUE){
 outputtables <- .PDE_extr_data_from_pdf(pdf = "/examples/Methotrexate/29973177_!.pdf",
 whattoextr = "tabandtxt",
 out = paste0(system.file(package = "PDE"),"/examples/MTX_output+-0_test/"),
 filter.words = strsplit("cohort;case-control;group;study population;study participants", ";")[[1]],
 ignore.case.fw = TRUE,
 regex.fw = FALSE,
 search.words = strsplit("(M|m)ethotrexate;(T|t)rexal;(R|r)heumatrex;(O|o)trexup", ";")[[1]],
 ignore.case.sw = FALSE,
 regex.sw = TRUE)
}

## Running an advanced analysis with filter and search words to
## extract sentences and tables and obtain documentation files
if(PDE_check_Xpdf_install() == TRUE){
 outputtables <- .PDE_extr_data_from_pdf(pdf = paste0(system.file(package = "PDE"),
                       "/examples/Methotrexate/29973177_!.pdf"),
 whattoextr = "tabandtxt",
 out = paste0(system.file(package = "PDE"),"/examples/MTX_output+-1_test/"),
 context = 1,
 dev_x = 20,
 dev_y = 9999,
 filter.words = strsplit("cohort;case-control;group;study population;study participants", ";")[[1]],
 ignore.case.fw = TRUE,
 regex.fw = FALSE,
 filter.word.times = 20,
 table.heading.words = "",
 ignore.case.th = FALSE,
 search.words = strsplit("(M|m)ethotrexate;(T|t)rexal;(R|r)heumatrex;(O|o)trexup", ";")[[1]],
 ignore.case.sw = FALSE,
 regex.sw = TRUE,
 eval.abbrevs = TRUE,
 out.table.format = ".csv (WINDOWS-1252)",
 write.table.locations = TRUE,
 write.tab.doc.file = TRUE,
 write.txt.doc.file = TRUE,
 exp.nondetc.tabs = TRUE,
 cpy_mv = "nocpymv",
 delete = TRUE)
}

}
\seealso{
\code{\link{PDE_pdfs2table}},\code{\link{PDE_pdfs2table_searchandfilter}},
\code{\link{PDE_pdfs2txt_searchandfilter}}

\code{\link{PDE_pdfs2table}},\code{\link{PDE_pdfs2table_searchandfilter}},
\code{\link{PDE_pdfs2txt_searchandfilter}}
}
