% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDE.R
\name{PDE-deprecated}
\alias{PDE-deprecated}
\title{Deprecated functions in package \sQuote{PDE}}
\description{
These functions are provided for compatibility with older versions
of \sQuote{PDE} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use
the replacement indicated below:
  \itemize{
    
    \item{PDE_install_XpdfReader4.02: \code{\link{PDE_install_Xpdftools4.02}}}
    
  }
}
