% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDE.R
\name{PDE_analyzer_i}
\alias{PDE_analyzer_i}
\title{Extracting data from PDF (Portable Document Format) files using a user interface}
\usage{
PDE_analyzer_i(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Indicates whether messages will be printed in the console. Default: \code{TRUE}.}
}
\description{
The \code{PDE_analyzer_i} provides a user interface for
the sentence and table extraction from multiple PDF files.
}
\section{Note}{

A detailed description of the elements in the user interface
can be found in the markdown file (README_PDE.md).
}

\examples{

PDE_analyzer_i()


}
