% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDE.R
\name{PDE_check_Xpdf_install}
\alias{PDE_check_Xpdf_install}
\title{Check if the XpdfReader is installed an in the system path}
\usage{
PDE_check_Xpdf_install()
}
\value{
The function returns a Boolean for the installation status and a message in case 
the commands are not detected.
}
\description{
\code{PDE_check_Xpdf_install} runs a version test for pdftotext, pdftohtml and pdftopng.
}
\examples{

PDE_check_Xpdf_install()

}
