% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDE.R
\name{PDE_install_XpdfReader4.02}
\alias{PDE_install_XpdfReader4.02}
\title{Install the XpdfReader 4.02}
\usage{
PDE_install_XpdfReader4.02(sysname = NULL, bin = NULL, verbose = TRUE)
}
\arguments{
\item{sysname}{String. In case the function returns "Unknown OS" the sysname can be set manually. 
Allowed options are "Windows", "Linux", and "Darwin" for Mac. Default: \code{NULL}.}

\item{bin}{String. In case the function returns "Unknown OS" the bin of the operational system
can be set manually. Allowed options are "64", and "32". Default: \code{NULL}.}

\item{verbose}{Logical. Indicates whether messages will be printed in the console. Default: \code{TRUE}.}
}
\description{
\code{PDE_download_XpdfReader4.02} is deprecated. Please run PDE_install_Xpdftools4.02() instead.
}
\examples{
\dontrun{

PDE_install_XpdfReader4.02()

}

}
