% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDE.R
\name{PDE_install_Xpdftools4.02}
\alias{PDE_install_Xpdftools4.02}
\title{Install the Xpdf command line tools 4.02}
\usage{
PDE_install_Xpdftools4.02(
  sysname = NULL,
  bin = NULL,
  verbose = TRUE,
  permission = 0
)
}
\arguments{
\item{sysname}{String. In case the function returns "Unknown OS" the sysname can be set manually. 
Allowed options are "Windows", "Linux", "SunOS" for Solaris, and "Darwin" for Mac. Default: \code{NULL}.}

\item{bin}{String. In case the function returns "Unknown OS" the bin of the operational system
can be set manually. Allowed options are "64", and "32". Default: \code{NULL}.}

\item{verbose}{Logical. Indicates whether messages will be printed in the console. Default: \code{TRUE}.}

\item{permission}{Numerical. If set to 0 the user is ask for a permission to
download Xpdftools. If set to 1, no user input is required. Default: \code{0}.}
}
\value{
The function returns a Boolean for the installation status and a message in case 
the commands are not installed.
}
\description{
\code{PDE_install_Xpdftools4.02} downloads and installs the XPDF command line tools 4.02.
}
\examples{
\dontrun{

PDE_install_Xpdftools4.02()

}


}
