% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDE.R
\name{PDE_reader_i}
\alias{PDE_reader_i}
\title{Browsing the PDE (PDF Data Extractor) analyzer results.}
\usage{
PDE_reader_i(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Indicates whether messages will be printed in the console. Default: \code{TRUE}.}
}
\description{
The \code{PDE_reader_i} allows the user-friendly visualization and quick-processing of the obtained results.
}
\section{Note}{

A detailed description of the elements in the user interface can be found in the markdown file (README_PDE.md)
}

\examples{
 
 PDE_reader_i()
 
}
