% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDE.R
\name{PDE_path}
\alias{PDE_path}
\title{Export the installation path the PDE (PDF Data Extractor) package}
\usage{
PDE_path()
}
\value{
The function returns a potential path for the PDE package. If the PDE
tool was not correctly installed it returns "".
}
\description{
\code{PDE_path} is deprecated. Please run system.file(package = "PDE") instead.
}
